account for empty results

+++ scripts/cluster/refresh-data-workers.sh
@@ -50,6 +50,11 @@ restart_data_workers() {
 
     local WORKER_COUNT=$(get_cluster_worker_count "$LOCAL_IP")
 
+    if [ "$WORKER_COUNT" -eq 0 ] || [ "$WORKER_COUNT" == "null" ] || [ -z "$WORKER_COUNT" ]; then
+        echo "No data workers found for $LOCAL_IP, skipping restart"
+        exit 0
+    fi
+
     echo "Found $WORKER_COUNT data workers for $LOCAL_IP, restarting..."
 
     sudo systemctl stop ${QUIL_DATA_WORKER_SERVICE_NAME}@{1..$WORKER_COUNT}
